requirements.txt
# Core
python-dotenv==1.0.1
PyYAML==6.0.1
loguru==0.7.2

# FastAPI / API Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Flask (dashboard fallback)
flask==3.0.3
flask-cors==4.0.1

# ML / Deep Learning
torch==2.2.2
torchvision==0.17.2
segmentation-models-pytorch==0.3.3
ultralytics==8.2.0
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.0
pillow==10.3.0
matplotlib==3.8.4
tqdm==4.66.4

# Geospatial
GDAL==3.8.4
rasterio==1.3.10
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
fiona==1.9.6
folium==0.16.0
branca==0.7.2

# Satellite Data Ingestion
pystac-client==0.7.7
planetary-computer==1.0.0
stackstac==0.5.1
asf-search==6.7.4

# Database
SQLAlchemy==2.0.30
GeoAlchemy2==0.15.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Alerts
twilio==9.0.4

# HTTP / Async
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.2

# Data handling
pandas==2.2.2
xarray==2024.3.0
dask==2024.4.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
